Start caching upload queries used as sub parts of other metrics again
