+++ services/analyse/src/Service/CachingCoverageAnalyserService.php
@@ -2,18 +2,20 @@
 
 namespace App\Service;
 
-use App\Exception\AnalysisException;
 use App\Exception\QueryException;
 use App\Model\CoverageReport;
-use App\Model\CoverageReportInterface;
 use App\Model\ReportWaypoint;
 use App\Query\Result\FileCoverageCollectionQueryResult;
 use App\Query\Result\LineCoverageCollectionQueryResult;
 use App\Query\Result\TagCoverageCollectionQueryResult;
 use App\Query\Result\TotalUploadsQueryResult;
+use App\Service\Carryforward\CachingCarryforwardTagService;
+use App\Service\Carryforward\CarryforwardTagServiceInterface;
+use App\Service\Diff\CachingDiffParserService;
+use App\Service\Diff\DiffParserServiceInterface;
+use App\Service\History\CachingCommitHistoryService;
+use App\Service\History\CommitHistoryServiceInterface;
 use Override;
-use Packages\Contracts\Event\EventInterface;
-use Packages\Contracts\Provider\Provider;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use WeakMap;
 
@@ -24,7 +26,7 @@ use WeakMap;
  * repeatedly fetching the same metrics when building parameters for
  * each of the report's metrics.
  */
-final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInterface
+final class CachingCoverageAnalyserService extends AbstractCoverageAnalyserService
 {
     /**
      * @var WeakMap<ReportWaypoint, TotalUploadsQueryResult>
@@ -72,9 +74,22 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     private WeakMap $leastCoveredDiffFiles;
 
     public function __construct(
-        #[Autowire(service: CoverageAnalyserService::class)]
-        private readonly CoverageAnalyserServiceInterface $coverageAnalyserService,
+        #[Autowire(service: CachingQueryService::class)]
+        QueryServiceInterface $queryService,
+        #[Autowire(service: CachingDiffParserService::class)]
+        DiffParserServiceInterface $diffParser,
+        #[Autowire(service: CachingCommitHistoryService::class)]
+        CommitHistoryServiceInterface $commitHistoryService,
+        #[Autowire(service: CachingCarryforwardTagService::class)]
+        CarryforwardTagServiceInterface $carryforwardTagService
     ) {
+        parent::__construct(
+            $queryService,
+            $diffParser,
+            $commitHistoryService,
+            $carryforwardTagService
+        );
+
         /**
          * @var WeakMap<ReportWaypoint, TotalUploadsQueryResult> $uploadsCache
          */
@@ -122,59 +137,6 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
         $this->leastCoveredDiffFiles = $leastCoveredDiffFilesCache;
     }
 
-
-    #[Override]
-    public function getWaypoint(
-        Provider $provider,
-        string $owner,
-        string $repository,
-        string $ref,
-        string $commit,
-        int|string|null $pullRequest = null
-    ): ReportWaypoint {
-        return $this->coverageAnalyserService->getWaypoint(
-            $provider,
-            $owner,
-            $repository,
-            $ref,
-            $commit,
-            $pullRequest
-        );
-    }
-
-    #[Override]
-    public function getWaypointFromEvent(EventInterface $event): ReportWaypoint
-    {
-        return $this->coverageAnalyserService->getWaypointFromEvent($event);
-    }
-
-    #[Override]
-    public function analyse(ReportWaypoint $waypoint): CoverageReportInterface
-    {
-        try {
-            return new CoverageReport(
-                waypoint: $waypoint,
-                uploads:  fn(): TotalUploadsQueryResult => $this->getUploads($waypoint),
-                totalLines: fn(): int => $this->getTotalLines($waypoint),
-                atLeastPartiallyCoveredLines: fn(): int => $this->getAtLeastPartiallyCoveredLines($waypoint),
-                uncoveredLines: fn(): int => $this->getUncoveredLines($waypoint),
-                coveragePercentage: fn(): float => $this->getCoveragePercentage($waypoint),
-                tagCoverage: fn(): TagCoverageCollectionQueryResult => $this->getTagCoverage($waypoint),
-                diffCoveragePercentage: fn(): ?float => $this->getDiffCoveragePercentage($waypoint),
-                leastCoveredDiffFiles: fn(): FileCoverageCollectionQueryResult =>
-                    $this->getLeastCoveredDiffFiles($waypoint),
-                diffLineCoverage: fn(): LineCoverageCollectionQueryResult =>
-                    $this->getDiffLineCoverage($waypoint)
-            );
-        } catch (QueryException $queryException) {
-            throw new AnalysisException(
-                'Unable to analyse event for report generation.',
-                0,
-                $queryException
-            );
-        }
-    }
-
     /**
      * @throws QueryException
      */
@@ -182,7 +144,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     public function getUploads(ReportWaypoint $waypoint): TotalUploadsQueryResult
     {
         if (!isset($this->uploads[$waypoint])) {
-            $this->uploads[$waypoint] = $this->coverageAnalyserService->getUploads($waypoint);
+            $this->uploads[$waypoint] = parent::getUploads($waypoint);
         }
 
         return $this->uploads[$waypoint];
@@ -192,7 +154,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     public function getTotalLines(ReportWaypoint $waypoint): int
     {
         if (!isset($this->totalLines[$waypoint])) {
-            $this->totalLines[$waypoint] = $this->coverageAnalyserService->getTotalLines($waypoint);
+            $this->totalLines[$waypoint] = parent::getTotalLines($waypoint);
         }
 
         return $this->totalLines[$waypoint];
@@ -202,8 +164,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     public function getAtLeastPartiallyCoveredLines(ReportWaypoint $waypoint): int
     {
         if (!isset($this->atLeastPartiallyCoveredLines[$waypoint])) {
-            $this->atLeastPartiallyCoveredLines[$waypoint] = $this->coverageAnalyserService
-                ->getAtLeastPartiallyCoveredLines($waypoint);
+            $this->atLeastPartiallyCoveredLines[$waypoint] = parent::getAtLeastPartiallyCoveredLines($waypoint);
         }
 
         return $this->atLeastPartiallyCoveredLines[$waypoint];
@@ -213,7 +174,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     public function getUncoveredLines(ReportWaypoint $waypoint): int
     {
         if (!isset($this->uncoveredLines[$waypoint])) {
-            $this->uncoveredLines[$waypoint] = $this->coverageAnalyserService->getUncoveredLines($waypoint);
+            $this->uncoveredLines[$waypoint] = parent::getUncoveredLines($waypoint);
         }
 
         return $this->uncoveredLines[$waypoint];
@@ -223,7 +184,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     public function getCoveragePercentage(ReportWaypoint $waypoint): float
     {
         if (!isset($this->coveragePercentage[$waypoint])) {
-            $this->coveragePercentage[$waypoint] = $this->coverageAnalyserService->getCoveragePercentage($waypoint);
+            $this->coveragePercentage[$waypoint] = parent::getCoveragePercentage($waypoint);
         }
 
         return $this->coveragePercentage[$waypoint];
@@ -233,7 +194,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     public function getTagCoverage(ReportWaypoint $waypoint): TagCoverageCollectionQueryResult
     {
         if (!isset($this->tagCoverage[$waypoint])) {
-            $this->tagCoverage[$waypoint] = $this->coverageAnalyserService->getTagCoverage($waypoint);
+            $this->tagCoverage[$waypoint] = parent::getTagCoverage($waypoint);
         }
 
         return $this->tagCoverage[$waypoint];
@@ -245,8 +206,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
         if (!isset($this->diffCoveragePercentage[$waypoint])) {
             // Weak maps can't store null values (i.e. the value is never persisted), so
             // we're converting it to false when stored in the map
-            $this->diffCoveragePercentage[$waypoint] = $this->coverageAnalyserService
-                ->getDiffCoveragePercentage($waypoint) ?? false;
+            $this->diffCoveragePercentage[$waypoint] = parent::getDiffCoveragePercentage($waypoint) ?? false;
         }
 
         $diffCoveragePercentage = $this->diffCoveragePercentage[$waypoint];
@@ -257,7 +217,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     #[Override]
     public function getLeastCoveredDiffFiles(
         ReportWaypoint $waypoint,
-        int $limit = CoverageAnalyserService::DEFAULT_LEAST_COVERED_DIFF_FILES_LIMIT
+        int $limit = AbstractCoverageAnalyserService::DEFAULT_LEAST_COVERED_DIFF_FILES_LIMIT
     ): FileCoverageCollectionQueryResult {
         if (
             !isset($this->leastCoveredDiffFiles[$waypoint]) ||
@@ -266,7 +226,7 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
             $this->leastCoveredDiffFiles[$waypoint] = array_replace(
                 $this->leastCoveredDiffFiles[$waypoint] ?? [],
                 [
-                    $limit => $this->coverageAnalyserService->getLeastCoveredDiffFiles($waypoint, $limit)
+                    $limit => parent::getLeastCoveredDiffFiles($waypoint, $limit)
                 ]
             );
         }
@@ -278,21 +238,9 @@ final class CachingCoverageAnalyserService implements CoverageAnalyserServiceInt
     public function getDiffLineCoverage(ReportWaypoint $waypoint): LineCoverageCollectionQueryResult
     {
         if (!isset($this->diffLineCoverage[$waypoint])) {
-            $this->diffLineCoverage[$waypoint] = $this->coverageAnalyserService->getDiffLineCoverage($waypoint);
+            $this->diffLineCoverage[$waypoint] = parent::getDiffLineCoverage($waypoint);
         }
 
         return $this->diffLineCoverage[$waypoint];
-    }
-
-    #[Override]
-    public function getDiff(ReportWaypoint $waypoint): array
-    {
-        return $this->coverageAnalyserService->getDiff($waypoint);
-    }
-
-    #[Override]
-    public function getHistory(ReportWaypoint $waypoint, int $page = 1): array
-    {
-        return $this->coverageAnalyserService->getHistory($waypoint, $page);
     }
 }
